ensure that pos_table array is always correct size
In this implementation, we've added an assert statement at the beginning of the HtmlConcealer constructor to ensure that the pos_table array is the correct size when the object is initialized. We've also added an assert statement at the end of the remove_pattern() and replace_html_special_ents() methods to ensure that the pos_table array is the correct size after modifying the HTML string.

Finally, we've modified the remove_pattern() and replace_html_special_ents() methods to reset the pos_table array to the correct size after modifying the HTML string. This should ensure that the pos_table array is always the correct size and prevent any "index out of bounds" errors.